+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Add explicit `#failure?` on `Response`. Return the opposite of `#success?`.
+* The `Command` handles the initialization process. Automatically provides
+getter and setter for each attributes sent to the command.
 
 ## 0.1.0 (2017-10-09)